Added support for assigning memory in CLMemoryProvider using float instead of double. Introduced CollectionFeatures::a and ExpressionComputation::generate. Replicated a common pattern for creating a kernel destination in DefaultCollectionEvaluable (this duplication of code is undesirable, but ultimately KernelizedEvaluable will be phased out completely so the functionality will eventually be consolidated). New and updated tests.

+++ utils/src/test/java/org/almostrealism/graph/model/test/PropagationTests.java
@@ -92,7 +92,7 @@ public class PropagationTests implements TestFeatures {
 		}
 
 		PackedCollection<?> input = new PackedCollection<>(size);
-		IntStream.range(0, size).forEach(i -> input.setMem(i, i));
+		IntStream.range(0, size).forEach(i -> input.setMem(i, (double) i));
 
 
 		PackedCollection<?> output;
